+++ modules/core/src/classes/quaternion.ts
@@ -31,7 +31,7 @@ import {
 } from '../gl-matrix/quat';
 // @ts-ignore gl-matrix types...
 import {transformQuat as vec4_transformQuat} from '../gl-matrix/vec4';
-import { Euler } from './euler';
+import type {Euler} from './euler';
 
 const IDENTITY_QUATERNION = [0, 0, 0, 1] as const;
 
@@ -84,28 +84,13 @@ export class Quaternion extends MathArray {
   }
 
   /**
-   * Creates a quaternion from the given Euler.
-   * @param euler
-   * @returns
+   * Sets this quaternion from Euler angles.
+   * @param euler - Euler angles and rotation order.
+   * @returns This quaternion.
    */
   fromEuler(euler: Euler): this {
-    this.identity();
-    switch (euler.order) {
-      case Euler.XYZ:
-        return this.rotateX(this[0]).rotateY(this[1]).rotateZ(this[2]);
-      case Euler.YXZ:
-        return this.rotateY(this[0]).rotateX(this[1]).rotateZ(this[2]);
-      case Euler.ZXY:
-        return this.rotateZ(this[0]).rotateX(this[1]).rotateY(this[2]);
-      case Euler.ZYX:
-        return this.rotateZ(this[0]).rotateY(this[1]).rotateX(this[2]);
-      case Euler.YZX:
-        return this.rotateY(this[0]).rotateZ(this[1]).rotateX(this[2]);
-      case Euler.XZY:
-        return this.rotateX(this[0]).rotateZ(this[1]).rotateY(this[2]);
-      default:
-        throw new Error('Unknown Euler angle order');
-    }
+    euler.getQuaternion(this);
+    return this;
   }
 
   fromAxisRotation(axis: Readonly<NumericArray>, rad: number): this {
@@ -309,10 +294,10 @@ export class Quaternion extends MathArray {
     arg0:
       | Readonly<NumericArray>
       | {
-        start: Readonly<NumericArray>;
-        target: Readonly<NumericArray>;
-        ratio: number;
-      },
+          start: Readonly<NumericArray>;
+          target: Readonly<NumericArray>;
+          ratio: number;
+        },
     arg1?: Readonly<NumericArray> | number,
     arg2?: number
   ): this {